+++ beat.h
@@ -4,6 +4,6 @@
 #include <stdbool.h>
 #include "./libs/btrack/BTrack_wrapper.h"
 
-bool Beat_isBeat(double* frame, int frameSize);
+bool Beat_isBeat(short* frame, int frameSize);
 // bool Beat_isBeat(double* frame, int frameSize, BTrack* bt);
 #endif
